+++ v3/crates/engine/tests/execution.rs
@@ -1351,6 +1351,43 @@ fn test_model_select_many_permission_filter_remote_relationships_array() -> anyh
     common::test_execution_expectation(test_path_string, &[common_metadata_path_string])
 }
 
+#[test]
+fn test_model_select_many_permission_filter_nested_select_object() -> anyhow::Result<()> {
+    common::test_execution_expectation_for_multiple_ndc_versions(
+        "execute/models/select_many/permission_filter/nested_select/object",
+        &["execute/models/select_many/permission_filter/nested_select/common-metadata.json"],
+        BTreeMap::from([
+            (
+                NdcVersion::V01,
+                vec!["execute/common_metadata/postgres_connector_ndc_v01_schema.json"],
+            ),
+            (
+                NdcVersion::V02,
+                vec!["execute/common_metadata/postgres_connector_ndc_v02_schema.json"],
+            ),
+        ]),
+    )
+}
+
+#[test]
+fn test_model_select_many_permission_filter_nested_relationships() -> anyhow::Result<()> {
+    common::test_execution_expectation_for_multiple_ndc_versions(
+        "execute/models/select_many/permission_filter/nested_relationships",
+        &[],
+        BTreeMap::from([
+            // This test can't use the old NDC v0.1.x connector, it does not support nested relationships in predicates
+            // (
+            //     NdcVersion::V01,
+            //     vec!["execute/common_metadata/custom_connector_v01_schema.json"],
+            // ),
+            (
+                NdcVersion::V02,
+                vec!["execute/common_metadata/custom_connector_v02_schema.json"],
+            ),
+        ]),
+    )
+}
+
 // ---------- Limit and Offset Tests
 #[test]
 fn test_model_select_many_limit_offset() -> anyhow::Result<()> {