+++ v3/crates/open-dds/src/types.rs
@@ -1,12 +1,5 @@
 use std::ops::Deref;
 
-use indexmap::IndexMap;
-use schemars::JsonSchema;
-use serde::{
-    de::value::{StrDeserializer, StringDeserializer},
-    Deserialize, Serialize,
-};
-
 use crate::commands::ArgumentMapping;
 use crate::{
     arguments::ArgumentName,
@@ -18,6 +11,12 @@ use crate::{
     impl_JsonSchema_with_OpenDd_for, impl_OpenDd_default_for,
     models::EnableAllOrSpecific,
     str_newtype,
+};
+use indexmap::IndexMap;
+use schemars::JsonSchema;
+use serde::{
+    de::value::{StrDeserializer, StringDeserializer},
+    Deserialize, Serialize,
 };
 
 #[derive(